+++ components/Panel/PanelFooter.jsx
@@ -6,16 +6,12 @@ class PanelFooter extends PureComponent {
 
   render() {
     const { prefixCls, className, title, more } = this.props;
-
-    const cls = classnames({
-      [`${prefixCls}-footer`]: true,
-      [className]: !!className,
-    });
+    const cls = classnames(`${prefixCls}-footer`, className);
 
     return (
       <div className={cls}>
-        { title && <div className={`${prefixCls}-title`}>{title}</div> }
-        { more && <div className={`${prefixCls}-more`}>{more}</div> }
+        {title && <div className={`${prefixCls}-title`}>{title}</div>}
+        {more && <div className={`${prefixCls}-more`}>{more}</div>}
       </div>
     );
   }