+++ components/SwipeAction/SwipeAction.jsx
@@ -13,9 +13,6 @@ class SwipeAction extends PureComponent {
     };
     this.isOpen = false;
     this.touchEnd = true;
-    this.onDragStart = this.onDragStart.bind(this);
-    this.onDragMove = this.onDragMove.bind(this);
-    this.onDragEnd = this.onDragEnd.bind(this);
   }
 
   componentDidMount() {
@@ -26,7 +23,7 @@ class SwipeAction extends PureComponent {
     Events.off(document.body, 'touchstart', e => this.onCloseSwipe(e));
   }
 
-  onDragStart() {
+  onDragStart = () => {
     if (this.isOpen) {
       this.touchEnd = false;
       this.close();
@@ -35,7 +32,7 @@ class SwipeAction extends PureComponent {
     this.touchEnd = true;
   }
 
-  onDragMove(event, { offsetX, offsetY }) {
+  onDragMove = (event, { offsetX, offsetY }) => {
     if (!this.touchEnd) return;
 
     const { disabled } = this.props;
@@ -56,12 +53,12 @@ class SwipeAction extends PureComponent {
 
     event.preventDefault();
 
-    this.doTransition({ offsetLeft: offsetX, duration: 0 });
+    this.doTransition({ offsetLeft: offsetX, animationDuration: 0 });
     return true;
   }
 
-  onDragEnd(event, { offsetX, startTime }) {
-    const { duration, moveDistanceRatio, moveTimeSpan } = this.props;
+  onDragEnd = (event, { offsetX, startTime }) => {
+    const { animationDuration, moveDistanceRatio, moveTimeSpan } = this.props;
     const timeSpan = new Date().getTime() - startTime.getTime();
     const btnsLeftWidth = this.left && this.left.offsetWidth;
     const btnsRightWidth = this.right && this.right.offsetWidth;
@@ -85,11 +82,11 @@ class SwipeAction extends PureComponent {
       this.close();
     } else {
       // 还原
-      this.doTransition({ offsetLeft: distanceX, duration });
+      this.doTransition({ offsetLeft: distanceX, animationDuration });
     }
   }
 
-  onBtnClick(e, btn) {
+  onBtnClick = (e, btn) => {
     e.preventDefault();
     const onClick = btn.onClick;
     if (onClick) {
@@ -101,7 +98,7 @@ class SwipeAction extends PureComponent {
     }
   }
 
-  onCloseSwipe(e) {
+  onCloseSwipe = (e) => {
     if (!this.wrap) return;
 
     if (this.isOpen) {
@@ -122,25 +119,25 @@ class SwipeAction extends PureComponent {
     }
   }
 
-  open(offsetLeft) {
-    const { duration, onOpen } = this.props;
+  open = (offsetLeft) => {
+    const { animationDuration, onOpen } = this.props;
     this.isOpen = true;
-    this.doTransition({ offsetLeft, duration });
+    this.doTransition({ offsetLeft, animationDuration });
     typeof onOpen === 'function' && onOpen();
   }
 
-  close() {
-    const { duration, onClose } = this.props;
+  close = () => {
+    const { animationDuration, onClose } = this.props;
     this.isOpen = false;
-    this.doTransition({ offsetLeft: 0, duration });
+    this.doTransition({ offsetLeft: 0, animationDuration });
     typeof onClose === 'function' && onClose();
   }
 
-  doTransition({ offsetLeft, duration }) {
-    this.setState({ offsetLeft, duration });
+  doTransition = ({ offsetLeft, animationDuration }) => {
+    this.setState({ offsetLeft, animationDuration });
   }
 
-  renderButtons(buttons, ref) {
+  renderButtons = (buttons, ref) => {
     const prefixCls = this.props.prefixCls;
 
     return (buttons && buttons.length) ? (
@@ -150,11 +147,8 @@ class SwipeAction extends PureComponent {
         {
           buttons.map((btn, i) => {
             const { theme, className, text } = btn;
-
-            const classes = classnames({
-              [`${prefixCls}-button`]: true,
+            const classes = classnames(`${prefixCls}-button`, className, {
               [`theme-${theme}`]: true,
-              [className]: !!className,
             });
 
             return (
@@ -173,23 +167,18 @@ class SwipeAction extends PureComponent {
 
   render() {
     const { prefixCls, className, left, right, children } = this.props;
-    const { offsetLeft, duration } = this.state;
-
-    const classes = classnames({
-      [`${prefixCls}`]: true,
-      [className]: !!className,
-    });
-
+    const { offsetLeft, animationDuration } = this.state;
+    const cls = classnames(`${prefixCls}`, className);
     const style = {
-      WebkitTransitionDuration: `${duration}ms`,
-      transitionDuration: `${duration}ms`,
+      WebkitTransitionDuration: `${animationDuration}ms`,
+      transitionDuration: `${animationDuration}ms`,
       WebkitTransform: `translate3d(${offsetLeft}px, 0, 0)`,
       transform: `translate3d(${offsetLeft}px, 0, 0)`,
     };
 
     return (left.length || right.length)
       ? (
-        <div className={classes} ref={(wrap) => { this.wrap = wrap; }}>
+        <div className={cls} ref={(wrap) => { this.wrap = wrap; }}>
           {this.renderButtons(left, 'left')}
           {this.renderButtons(right, 'right')}
           <Drag
@@ -213,7 +202,7 @@ SwipeAction.propTypes = {
   right: PropTypes.arrayOf(PropTypes.object),
   moveDistanceRatio: PropTypes.number,
   moveTimeSpan: PropTypes.number,
-  duration: PropTypes.number,
+  animationDuration: PropTypes.number,
   offset: PropTypes.number,
   onOpen: PropTypes.func,
   onClose: PropTypes.func,
@@ -225,7 +214,7 @@ SwipeAction.defaultProps = {
   right: [],
   moveDistanceRatio: 0.5,
   moveTimeSpan: 300,
-  duration: 300,
+  animationDuration: 300,
   offset: 10,
 };