+++ examples/pages/SwitchPage.jsx
@@ -9,13 +9,13 @@ class Page extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      switch: false,
+      value: false,
     };
   }
 
-  toggle(value) {
+  toggle = (value) => {
     this.setState({
-      switch: value,
+      value,
     });
   }
 
@@ -38,13 +38,6 @@ class Page extends Component {
               <Cell description={<Switch defaultChecked />}>默认开</Cell>
               <Cell description={<Switch disabled />}>禁用的开关（默认关）</Cell>
               <Cell description={<Switch defaultChecked disabled />}>禁用的开关（默认开）</Cell>
-            </Panel.Body>
-          </Panel>
-          <Panel>
-            <Panel.Header title="受控开关" />
-            <Panel.Body>
-              <Cell description={<Switch onChange={value => this.toggle(value)} />}>主开关</Cell>
-              <Cell description={<Switch checked={this.state.switch} onChange={value => console.log(value)} />}>受控开关</Cell>
             </Panel.Body>
           </Panel>
         </main>