+++ scripts/webpack/config.umd.js
@@ -4,8 +4,9 @@ const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 // const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 const config = require('./config.base');
+const version = require('../../package.json').version;
 
-const env = process.env.WEBPACK_ENV;
+const env = process.env.NODE_ENV;
 
 config.devtool = 'source-map';
 
@@ -40,7 +41,6 @@ config.externals = {
 
 const cssConfig = {
   filename: '[name].css',
-  disable: false,
   allChunks: true,
 };
 
@@ -69,20 +69,28 @@ if (env === 'production') {
     sourceMap: true,
     mangle: true,
   }));
+
+  config.plugins.push(new webpack.BannerPlugin(`
+  Zarm v${version}
+
+  Github: https://github.com/ZhonganTechENG/zarm
+ 
+  Copyright (c) 2013-present, ZhonganTech, Inc.
+ 
+  This source code is licensed under the MIT license found in the
+  LICENSE file in the root directory of this source tree.
+  `));
+
+  // config.plugins.push(new BundleAnalyzerPlugin({
+  //   analyzerMode: 'static',
+  // }));
 } else {
   config.plugins.push(new ExtractTextPlugin(cssConfig));
 }
 
 config.plugins.push(new webpack.DefinePlugin({
-  'process.env': {
-    NODE_ENV: '"production"',
-  },
+  'process.env.NODE_ENV': JSON.stringify(env || 'production'),
   __DEBUG__: false,
 }));
-
-
-// config.plugins.push(new BundleAnalyzerPlugin({
-//   analyzerMode: 'static'
-// }));
 
 module.exports = config;